Fix: remove updateLine API's response body
