+++ src/main/java/subway/controller/dto/LineUpdateRequest.java
@@ -1,0 +1,42 @@
+package subway.controller.dto;
+
+public class LineUpdateRequest {
+
+    private final String name;
+
+    private final String color;
+
+    private final Long upStationId;
+
+    private final Long downStationId;
+
+    private final Long distance;
+
+    public LineUpdateRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
+        this.name = name;
+        this.color = color;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public Long getDistance() {
+        return distance;
+    }
+}
